requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
click==8.2.1
Flask==3.1.1
frozenlist==1.7.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.4
propcache==0.3.2
pyee==11.1.1
python-dotenv==1.1.1
slackclient==2.9.4
slackeventsapi==3.0.3
typing_extensions==4.14.0
Werkzeug==3.1.3
yarl==1.20.1
